translation(newsletter): 47
* German Translation Newsletter 47 fin4dao

Hi, I converted the $ amounts into € too. Please change back if you don't like it or DM on discord <fin4thepeople#8470> so I can change it.

Thank you!

* Updated German Translation Newsletter 47 fin4dao

Added:
Line 7: Keep the currency and all currencies in $ as in the original text.
Line 15:Account should be changed to Konto

Next translations will be informal

Thank you for proof reading Nesy and thanks for explaining how to change the PR Dark